7 implement remaining data flow

+++ WA.Pizza.Tests/TestDatabaseFixture.cs
@@ -9,32 +9,34 @@ using Xunit;
 
 namespace WA.Pizza.Tests
 {
-    public class TestDatabaseFixture
+    public class TestDatabaseFixture: IDisposable
     {
-
-    }
-    [CollectionDefinition("Test database collection")]
-    public class DatabaseCollection : ICollectionFixture<TestDatabaseFixture>,  IDisposable
-    {
-        public ApplicationDbContext applicationDbContext { get; private set; }
-        public DatabaseCollection()
+        ApplicationDbContext applicationDbContext;
+        public TestDatabaseFixture()
         {
+
             var config = new ConfigurationBuilder()
-                               .AddJsonFile("appsettings_test.json")
-                               .Build();
+                       .AddJsonFile("appsettings_test.json")
+                       .Build();
             var optionsBuilder = new DbContextOptionsBuilder<ApplicationDbContext>()
                .UseSqlServer(config.GetConnectionString("Test"));
             applicationDbContext = new ApplicationDbContext(optionsBuilder.Options);
-
-            applicationDbContext.Database.EnsureDeleted();
+            if (applicationDbContext.Database.CanConnect())
+                applicationDbContext.Database.EnsureDeleted();
             applicationDbContext.Database.Migrate();
 
         }
 
-        public void Dispose()
+        void IDisposable.Dispose()
         {
-            applicationDbContext.Database.EnsureDeleted();
+            if (applicationDbContext.Database.CanConnect())
+                applicationDbContext.Database.EnsureDeleted();
         }
+    }
+    [CollectionDefinition("Test database collection")]
+    public class DatabaseCollection : ICollectionFixture<TestDatabaseFixture>
+    {
+
 
         // This class has no code, and is never created. Its purpose is simply
         // to be the place to apply [CollectionDefinition] and all the
